auth: remove oauth flux implementation
